удалил класс Tasks, добавил AvailableActions
